# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AdminPayment {
  id: string;
  telegramUserId: string;
  /** @nullable */
  username?: string | null;
  tariffId: string;
  amount: number;
  status: string;
}

export interface AdminStats {
  totalUsers: number;
  totalRevenue: number;
  activeSubscriptions: number;
  expiredSubscriptions: number;
  trialUsers: number;
  pendingPayments: number;
  paidPayments: number;
  revenueToday: number;
  newUsersToday: number;
  recentPayments: AdminPayment[];
}

export interface AdminUser {
  id: string;
  /** @nullable */
  username?: string | null;
  balance: number;
  hasUsedTrial: boolean;
  /** @nullable */
  referredById?: string | null;
  subscriptionCount: number;
}

export interface AdminUserList {
  items: AdminUser[];
  total: number;
  page: number;
  limit: number;
}

export interface AdminPaymentList {
  items: AdminPayment[];
  total: number;
  page: number;
  limit: number;
}

export interface AdminSubscription {
  id: string;
  telegramUserId: string;
  /** @nullable */
  username?: string | null;
  vpnKey?: string;
  tariffId: string;
  expiresAt: string;
  isActive: boolean;
}

export interface AdminSubscriptionList {
  items: AdminSubscription[];
  total: number;
  page: number;
  limit: number;
}

export interface AdminPromocode {
  id: string;
  bonusAmount: number;
  maxUses: number;
  usesCount: number;
}

export interface PromocodeInput {
  id: string;
  bonusAmount: number;
  maxUses: number;
}

export interface AuthConfig {
  botUsername: string;
}

export interface TelegramAuthData {
  id: number;
  hash: string;
  auth_date: number;
  first_name?: string;
  last_name?: string;
  username?: string;
  photo_url?: string;
}

export interface AuthUser {
  id: number;
  firstName: string;
  username?: string;
  lastName?: string;
}

export interface ResellerProfile {
  /** Reseller balance in rubles */
  balance: number;
  /** Current wholesale discount percentage (0-40) */
  discount: number;
}

export interface ResellerClient {
  id: string;
  vpnKey: string;
  expiresAt: string;
  tariffId: string;
}

export interface CreateResellerClientBody {
  tariffId: string;
  /** Optional label/name for the client */
  label?: string;
}

export interface RenewSubscriptionBody {
  tariffId: string;
}

export interface BalanceAdjust {
  delta: number;
}

export type ListAdminUsersParams = {
page?: number;
limit?: number;
search?: string;
};

export type ListAdminPaymentsParams = {
page?: number;
limit?: number;
status?: ListAdminPaymentsStatus;
};

export type ListAdminPaymentsStatus = typeof ListAdminPaymentsStatus[keyof typeof ListAdminPaymentsStatus];


export const ListAdminPaymentsStatus = {
  pending: 'pending',
  paid: 'paid',
  failed: 'failed',
} as const;

export type ListAdminSubscriptionsParams = {
page?: number;
limit?: number;
active?: boolean;
};

export type Logout200 = {
  ok: boolean;
};

export type DeleteSubscription200 = {
  ok: boolean;
};